% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_legend_sr.R
\name{add_legend_sr}
\alias{add_legend_sr}
\title{Add a legend to a Leaflet map}
\usage{
add_legend_sr(map, colors, labels, sizes, opacity = 1, group)
}
\arguments{
\item{map}{A Leaflet map object}

\item{colors}{A string of colours, e.g. \code{c("black", "yellow", "yellow")}.}

\item{labels}{A list of labels, e.g. \code{c(0:10)}.}

\item{sizes}{A list of sizes, e.g. \code{c(5, 20, 40)}.}

\item{opacity}{The legend's opacity as decimal number between 0 and 1,
default: \code{1}.}

\item{group}{A string containing the intended layer name,
e.g. \code{"Species richness"}.}
}
\value{
The output of \code{leaflet::addLegend}.
}
\description{
This function wraps \code{leaflet::addLegend} with some defaults.
}
