% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_legend_ta.R
\name{add_legend_ta}
\alias{add_legend_ta}
\title{Add a legend to a Leaflet map}
\usage{
add_legend_ta(map, colors, labels, sizes, opacity = 1, group)
}
\arguments{
\item{map}{A Leaflet map object}

\item{colors}{A string of colours, e.g. `c("black", "yellow", "yellow")`.}

\item{labels}{A list of labels, e.g. `c(0:10)`.}

\item{sizes}{A list of sizes, e.g. `c(5, 20, 40)`.}

\item{opacity}{The legend's opacity as decimal number between 0 and 1,
default: `1`.}

\item{group}{A string containing the intended layer name,
e.g. `"Total abundance"`.}
}
\value{
The output of `leaflet::addLegend`.
}
\description{
This function wraps `leaflet::addLegend` with some defaults.
}
