% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate_data.R
\name{consolidate_data}
\alias{consolidate_data}
\title{Consolidate all data to a single set of files}
\usage{
consolidate_data(dir)
}
\arguments{
\item{dir}{File path for the iLab dashboard}
}
\value{
Saves data
}
\description{
Consolidate all data to a single set of files
}
