% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_data.R
\name{copy_data}
\alias{copy_data}
\title{Copy data to use within DBCA sharepoint folders}
\usage{
copy_data(dir = iLab::get_dir("iLab_fish"))
}
\arguments{
\item{dir}{File path for the iLab}
}
\value{
Copies data into folders
}
\description{
Copy data to use within DBCA sharepoint folders
}
