% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{corals}
\alias{corals}
\title{Some completely made up coral data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 65100 rows and 14 columns.
}
\usage{
corals
}
\description{
Some completely made up coral data
}
\seealso{
Other included: 
\code{\link{fish}},
\code{\link{print.mpa_data}()}
}
\concept{included}
\keyword{datasets}
