% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_source_data.R
\name{download_source_data}
\alias{download_source_data}
\title{Download source data files from CKAN to the local data directory}
\usage{
download_source_data(
  ckan_url = Sys.getenv("CKAN_URL", unset = "https://data.dbca.wa.gov.au"),
  data_dir = here::here("inst/data")
)
}
\arguments{
\item{ckan_url}{The data catalogue URL,
default: \code{Sys.getenv("CKAN_URL")} or the current DBCA CKAN server.}

\item{data_dir}{A local directory to download files to,
default: \code{here::here("inst/data")}.}
}
\value{
None.
}
\description{
This method can be configured to point to the data file through environment
variables. It will work with defaults unless the location of the data file
changes.
}
