% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fish}
\alias{fish}
\title{Some completely made up fish data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 9609 rows and 32 columns.
}
\usage{
fish
}
\description{
Some completely made up fish data
}
\seealso{
Other included: 
\code{\link{corals}}
}
\concept{included}
\keyword{datasets}
