% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate one data object to use in server.R}
\usage{
generate_data(save = TRUE, dest = here::here("inst/data/mpa_data.rds"))
}
\arguments{
\item{save}{Boolean whether to save data to rds}

\item{dest}{The target file to save rds as,
default: here::here("inst/data/mpa_data.rds")}
}
\value{
An object of class "mpa_data" containing all data tibbles and objects
  used by server.R
}
\description{
Generate one data object to use in server.R
}
\examples{
\dontrun{
x <- generate_data(save=FALSE) # only returns data
x <- generate_data() # returns data and saves data to local file
}
}
