% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plots.R
\name{generate_plots}
\alias{generate_plots}
\title{Generate one data object to use in server.R}
\usage{
generate_plots()
}
\value{
All plots
}
\description{
Generate one data object to use in server.R
}
\examples{
\dontrun{
x <- generate_plots()
}
}
