% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_mpatheme.R
\name{ggplot_mpatheme}
\alias{ggplot_mpatheme}
\title{A custom GGplot2 theme for mpaviewer}
\usage{
ggplot_mpatheme(n_years = NULL)
}
\arguments{
\item{n_years}{Integer value used to scale x-axis text size. Defaults to NULL.}
}
\value{
A custom ggplot2::theme()
}
\description{
A custom GGplot2 theme for mpaviewer
}
