% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googledrive_download_data.R
\name{googledrive_download_data}
\alias{googledrive_download_data}
\title{Download the application data file mpa_data.rds from CKAN}
\usage{
googledrive_download_data(raw_dir, project_dir)
}
\arguments{
\item{raw_dir}{A local directory to download files to.}

\item{project_dir}{A local directory containing the Github repository.}
}
\value{
None.
}
\description{
This method can be configured to point to the data file through environment
variables. It will work with defaults unless the location of the data file
changes.
}
