% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpaviewer-package.R, R/mpaviewer.R
\docType{package}
\name{mpaviewer-package}
\alias{mpaviewer}
\alias{mpaviewer-package}
\title{mpaviewer: MPA Viewer}
\description{
An interactive overview of selected biodiversity assets in Marine Protected Areas of Western Australia.

\code{mpaviewer} is an interactive dashboard for
Western Australian marine monitoring data,
hosted at \url{https://mpaviewer.dbca.wa.gov.au/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dbca-wa/mpaviewer}
  \item Report bugs at \url{https://github.com/dbca-wa/mpaviewer/issues}
}


Useful links:
\itemize{
  \item \url{https://github.com/dbca-wa/mpaviewer}
  \item Report bugs at \url{https://github.com/dbca-wa/mpaviewer/issues}
}

}
\author{
\strong{Maintainer}: Brooke Gibbons \email{brooke.gibbons@uwa.edu.au}

}
\keyword{internal}
