% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{print.mpa_data}
\alias{print.mpa_data}
\title{S3 print method for 'mpa_data'.}
\usage{
\method{print}{mpa_data}(x, ...)
}
\arguments{
\item{x}{An object of class \code{mpa_data} as returned by \code{generate_data}.}

\item{...}{Extra parameters for \code{print}}
}
\description{
Prints a short representation of mpa_data returned by
\code{generate_data}.
}
\concept{included}
