% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dbca_files_csv.R
\name{read_dbca_files_csv}
\alias{read_dbca_files_csv}
\title{Read DBCA CSV files}
\usage{
read_dbca_files_csv(flnm)
}
\arguments{
\item{flnm}{A filename}
}
\value{
The file contents as tibble
}
\description{
Read DBCA CSV files
}
