% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dbca_files_txt.R
\name{read_dbca_files_txt}
\alias{read_dbca_files_txt}
\title{Read DBCA TXT files}
\usage{
read_dbca_files_txt(flnm, data_dir)
}
\arguments{
\item{flnm}{A filename}

\item{data_dir}{The data directory, default: \code{here::here("inst/data")}.}
}
\value{
The file contents as tibble
}
\description{
Read DBCA TXT files
}
