% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.max.R
\name{se.max}
\alias{se.max}
\title{Function for calculating standard error maximum}
\usage{
se.max(x)
}
\arguments{
\item{x}{}
}
\value{
standard error maxmimum
}
\description{
Function for calculating standard error maximum
}
