% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.min.R
\name{se.min}
\alias{se.min}
\title{Function for calculating standard error minimum}
\usage{
se.min(x)
}
\arguments{
\item{x}{}
}
\value{
standard error minimum
}
\description{
Function for calculating standard error minimum
}
