% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Donwload all files from a Google Drive folder to a local folder}
\usage{
download_data(
  drive_folder = Sys.getenv("GD_MPAVIEWER"),
  api_key = Sys.getenv("GOOGLE_API_KEY"),
  data_dir = here::here("inst/data")
)
}
\arguments{
\item{drive_folder}{The URL of the Google Drive folder,
default: \code{Sys.getenv("GD_MPAVIEWER")}}

\item{api_key}{A read-permitted Google API key,
default: \code{Sys.getenv("GOOGLE_API_KEY")}.
Note the Google Drive API must be enabled.}

\item{data_dir}{A local directory to download files to,
default: \code{here::here("inst/data")}.}
}
\value{
None.
}
\description{
NOTE this function does not work yet
}
