% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Download the data file from CKAN}
\usage{
download_data(
  ckan_url = Sys.getenv("CKAN_URL"),
  data_rid = Sys.getenv("MPAVIEWER_RID"),
  data_dir = here::here("inst/data")
)
}
\arguments{
\item{ckan_url}{The data catalogue URL,
default: \code{Sys.getenv("CKAN_URL")}}

\item{data_rid}{The resource ID of the data file,
default: \code{Sys.getenv("MPAVIEWER_RID")}.}

\item{data_dir}{A local directory to download files to,
default: \code{here::here("inst/data")}.}
}
\value{
None.
}
\description{
Download the data file from CKAN
}
