% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plots.R
\name{generate_plots}
\alias{generate_plots}
\title{Generate plots to display in the dashboard}
\usage{
generate_plots()
}
\value{
All plots
}
\description{
Generate plots to display in the dashboard
}
\examples{
\dontrun{
x <- generate_plots()
}
}
