% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpaviewer.R
\docType{package}
\name{mpaviewer-package}
\alias{mpaviewer}
\alias{mpaviewer-package}
\title{mpaviewer: MPA Viewer}
\description{
\code{mpaviewer} is an interactive dashboard for
Western Australian marine monitoring data,
hosted at \url{https://mpaviewer.dbca.wa.gov.au/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dbca-wa/mpaviewer}
  \item \url{https://dbca-wa.github.io/mpaviewer/}
  \item Report bugs at \url{https://github.com/dbca-wa/mpaviewer/issues}
}

}
\author{
\strong{Maintainer}: Florian Mayer \email{Florian.Mayer@dbca.wa.gov.au}

}
\keyword{internal}
