% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpaviewer.R
\docType{package}
\name{mpaviewer-package}
\alias{mpaviewer}
\alias{mpaviewer-package}
\title{mpaviewer: MPA Viewer}
\description{
\code{mpaviewer} is an interactive dashboard for
Western Australian marine monitoring data,
hosted at \url{https://mpaviewer.dbca.wa.gov.au/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dbca-wa/mpaviewer}
  \item Report bugs at \url{https://github.com/dbca-wa/mpaviewer/issues}
}

}
\author{
\strong{Maintainer}: Florian W. Mayer \email{Florian.Mayer@dbca.wa.gov.au} (\href{https://orcid.org/0000-0003-4269-4242}{ORCID})

Authors:
\itemize{
  \item Brooke Gibbons \email{Brooke.Gibbons@uwa.edu.au} (\href{https://orcid.org/0000-0002-0628-7781}{ORCID})
}

Other contributors:
\itemize{
  \item DBCA [copyright holder, funder]
  \item WAMMP [funder]
}

}
\keyword{internal}
